dom/canvas/test/test_offscreencanvas_toblob.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebGL in OffscreenCanvas</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="offscreencanvas.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<canvas id="c-ref" width="64" height="64"></canvas>
<script>

SimpleTest.waitForExplicitFinish();

function createCanvas() {
  var htmlCanvas = document.createElement('canvas');
  htmlCanvas.width = 64;
  htmlCanvas.height = 64;
  document.body.appendChild(htmlCanvas);
  return htmlCanvas;
}

function TransferCtrlToOffscreen(canvas) {
  ok(canvas, "Should have HTML canvas element");
  ok(canvas.transferControlToOffscreen,
     "HTMLCanvasElement has transferControlToOffscreen function");
  return canvas.transferControlToOffscreen();
}

function testBlob(blob, callback) {
  // testing toBlob
  // Fill c-ref with green color.
  var c = document.getElementById("c-ref");
  var ctx = c.getContext("2d");
  ctx.rect(0, 0, 64, 64);
  ctx.fillStyle = "#00FF00";
  ctx.fill();
  var reader = new FileReader();
  reader.onload = function(e) {
    ok(c.toDataURL() == e.target.result,
       "toBlob should return a 64x64 green square");
    callback();
  };
  reader.readAsDataURL(blob);
}

function runTestOnMainThread() {

  window.onmessage = function(evt) {
    var msg = evt.data || {};
    if (msg.type == "test") {
      ok(msg.result, msg.name);
    }
    if (msg.type == "blob") {
      testBlob(msg.blob, SimpleTest.finish);
    }
  }

  var offscreenCanvas = TransferCtrlToOffscreen(createCanvas());
  ok(offscreenCanvas, "Expected transferControlToOffscreen to succeed");
  entryFunction('webgl', 'toblob', '', offscreenCanvas);
}

function runTest() {

  var worker = new Worker("offscreencanvas.js");

  worker.onmessage = function(evt) {
    var msg = evt.data || {};
    if (msg.type == "test") {
      ok(msg.result, msg.name);
    }
    if (msg.type == "blob") {
      testBlob(msg.blob, function() {
        if (msg.canvasType == "webgl") {
          worker.terminate();
          runTestOnMainThread();
        } else {
          // Start to test toblob for webgl
          offscreenCanvas = TransferCtrlToOffscreen(createCanvas());
          worker.postMessage({canvasType: 'webgl', testType: 'toblob',
                              canvas: offscreenCanvas}, [offscreenCanvas]);
        }
      });
    }
  }

  ok(worker, "Web worker successfully created");

  var offscreenCanvas = TransferCtrlToOffscreen(createCanvas());
  ok(offscreenCanvas, "Expected transferControlToOffscreen to succeed");

  // Start to test toblob for 2d
  worker.postMessage({canvasType: '2d', testType: 'toblob',
                      canvas: offscreenCanvas}, [offscreenCanvas]);
}

SpecialPowers.pushPrefEnv({'set': [
  ['gfx.offscreencanvas.enabled', true],
  ['webgl.force-enabled', true],
]}, runTest);

</script>
</body>
</html>
